rename {HIRT_Multichannel_Convolution => convolution}/
