--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "files.associations": {
-        "algorithm": "cpp"
-    }
-}
