+++ include/convolution/mono.hpp
