+++ include/convolution/multichannel.hpp
