+++ include/convolution/n_to_mono.hpp
