+++ include/convolution/partitioned.hpp
