+++ include/convolution/utilities.hpp
